# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Feature Voting System API
  version: 0.2.0
  description: >
    Current API contract for authentication, current-user, and feature voting
    endpoints.

servers:
  - url: /api/v1
    description: API v1 base path

tags:
  - name: Authentication
    description: User signup, login, and token refresh endpoints
  - name: Users
    description: Authenticated user profile endpoints
  - name: Features
    description: Feature creation, listing, and vote management endpoints

paths:
  /auth/signup/:
    post:
      tags:
        - Authentication
      operationId: authSignup
      summary: Sign up a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created and tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /auth/login/:
    post:
      tags:
        - Authentication
      operationId: authLogin
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User authenticated and tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /auth/refresh/:
    post:
      tags:
        - Authentication
      operationId: authRefresh
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /users/me/:
    get:
      tags:
        - Users
      operationId: currentUser
      summary: Get the current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /features/:
    get:
      tags:
        - Features
      operationId: featureList
      summary: List features
      description: >
        Public endpoint for listing features with optional search, sorting,
        and authenticated-user filters.
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Case-insensitive search across title and description.
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - created_at
              - last_voted_at
              - number_of_votes
            default: number_of_votes
          description: All sort options are returned in descending order.
        - in: query
          name: mine
          schema:
            type: boolean
            default: false
          description: Only include features owned by the authenticated user.
        - in: query
          name: voted
          schema:
            type: boolean
            default: false
          description: Only include features the authenticated user has voted on.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number for paginated results.
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Number of features per page. Defaults to 20.
      responses:
        "200":
          description: Paginated feature list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFeatureList"
        "401":
          description: Authentication required for mine or voted filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
    post:
      tags:
        - Features
      operationId: featureCreate
      summary: Create a feature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureRequest"
      responses:
        "201":
          description: Feature created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /features/{feature_id}/vote/:
    post:
      tags:
        - Features
      operationId: featureVote
      summary: Vote on a feature
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: feature_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote applied and feature counters updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "400":
          description: Invalid vote state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "404":
          description: Feature not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /features/{feature_id}/unvote/:
    post:
      tags:
        - Features
      operationId: featureUnvote
      summary: Remove a vote from a feature
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: feature_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote removed and feature counters updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feature"
        "400":
          description: Invalid vote state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "404":
          description: Feature not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignupRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
      additionalProperties: false

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          format: password
      additionalProperties: false

    RefreshTokenRequest:
      type: object
      required:
        - refresh
      properties:
        refresh:
          type: string
      additionalProperties: false

    TokenPair:
      type: object
      required:
        - access
        - refresh
      properties:
        access:
          type: string
          description: JWT access token with a 5-minute lifetime.
        refresh:
          type: string
          description: JWT refresh token with a 30-day lifetime.
      additionalProperties: false

    AccessToken:
      type: object
      required:
        - access
      properties:
        access:
          type: string
          description: JWT access token with a 5-minute lifetime.
      additionalProperties: false

    CurrentUser:
      type: object
      required:
        - id
        - username
      properties:
        id:
          type: integer
        username:
          type: string
      additionalProperties: false

    CreateFeatureRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
      additionalProperties: false

    Feature:
      type: object
      required:
        - id
        - title
        - description
        - owner
        - created_at
        - updated_at
        - number_of_votes
        - is_owner
        - has_voted
      properties:
        id:
          type: integer
        title:
          type: string
          maxLength: 255
        description:
          type: string
        owner:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_voted_at:
          type: string
          format: date-time
          nullable: true
        number_of_votes:
          type: integer
          minimum: 0
        is_owner:
          type: boolean
        has_voted:
          type: boolean
      additionalProperties: false

    PaginatedFeatureList:
      type: object
      required:
        - count
        - next
        - previous
        - results
      properties:
        count:
          type: integer
          minimum: 0
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
      additionalProperties: false

    ErrorDetail:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
      additionalProperties: false

    ValidationError:
      type: object
      additionalProperties:
        type: array
        items:
          type: string